fixup! Remove some utilities from the core to provider interface
